Rewrite unexpected-markdown to remove its duplicate runtime.
As evaldown evolved and the docs started to rely on its new behaviour
problems arose with unexpected-markdown.In particular, this module
attempted to replicate the behaviour of snippet evaluation and was
effectively a separate implementation of it which meant it became
out of sync.

In order to avoid this situation going forward, rewrite the module
to make use of the same runtime and merely assert that afterwards
the evaluated result matches the state of the on-disk markdown file.